share singletonkey and init table
