+++ services/user/CommonSys/include/services/system/CommonTables.hpp
@@ -1,0 +1,13 @@
+#pragma once
+#include <psibase/Table.hpp>
+
+namespace UserService
+{
+   struct InitializedRecord
+   {
+      psibase::SingletonKey key;
+   };
+   PSIO_REFLECT(InitializedRecord, key);
+   using InitTable = psibase::Table<InitializedRecord, &InitializedRecord::key>;
+
+}  // namespace UserService
