+++ services/user/NftSys/include/services/user/nftTables.hpp
@@ -8,19 +8,6 @@ namespace UserService
 {
    using NID = uint32_t;
 
-   // This structure is shared between several User services that need a
-   //   flag to track initialization
-   struct SingletonKey
-   {
-   };
-   PSIO_REFLECT(SingletonKey);
-   struct InitializedRecord
-   {
-      SingletonKey key;
-   };
-   PSIO_REFLECT(InitializedRecord, key);
-   using InitTable = psibase::Table<InitializedRecord, &InitializedRecord::key>;
-
    struct NftHolderRecord
    {
       psibase::AccountNumber account;
